bench/requirements.txt
memory-profiler
pytest-benchmark
pytest-random-order
python-rapidjson
simplejson
tabulate
ujson
dynamodb_json
